% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{convert_datcub_to_datmat}
\alias{convert_datcub_to_datmat}
\title{Convert a data cube to a data matrix}
\usage{
convert_datcub_to_datmat(datcub)
}
\arguments{
\item{datcub}{either a matrix or 3d array}
}
\value{
one of:
\itemize{
   \item a vector (if datcub is a matrix) of length = nrow(datcub)*ncol(datcub)
   \item a matrix (if datcub is an array) of dimension (nrow(datcub)*ncol(datcub), nbands(datcub))
}


Example:
convert_datcub_to_datmat(read_khoros_viff('8class-caseI-n_m0v5.viff'))
}
\description{
Converts two or three dimensional arrays in [row,col] layout (such as those returned from \code{read_khoros_viff})
to either a vector (if cube only has 1 band) or matrix (if cube has > 1).
The columns of the data matrix correspond to the bands of the data cube.
}
\details{
The rows of the data matrix are populated by the pixels of the cube, starting from the (x=1,y=1) position of the cube, then
(x=2,y=1), ..., (x=img_x,y=1), (x=1, y=2), ..., (x=img_x, y=img_y)
}
