% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{convert_datmat_to_datcub}
\alias{convert_datmat_to_datcub}
\title{Convert a data matrix to a data cube}
\usage{
convert_datmat_to_datcub(datmat, img_x, img_y)
}
\arguments{
\item{datmat}{a data matrix of dimension (img_x*img_y, num bands)}

\item{img_x}{an integer specifying the width of the image in pixels}

\item{img_y}{an integer specifying the height of the image in pixels}
}
\value{
a 3d array of dimension (img_y, img_x, num bands).  The first row of X will be put in the [1,1,] entry of the result.

Example:
convert_datmat_to_datcub(convert_datcub_to_datmat(read_khoros_viff('8class-caseI-n_m0v5.viff')))
}
\description{
The rows of a data matrix contain the pixel spectra.
This function puts a data matrix back into a cube (normally 3d array, 2d matrix if there is only one band)
}
