% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{convert_linidx_to_pxlidx}
\alias{convert_linidx_to_pxlidx}
\title{Convert linear indices to pixel indices
.}
\usage{
convert_linidx_to_pxlidx(linidx, img_x, img_y)
}
\arguments{
\item{linidx}{linear index(ex) to convert}

\item{img_x}{the width of the image in pixels}

\item{img_y}{the height of the image in pixels}
}
\value{
a two-column matrix of [row,col] where row in 1:img_y and col in 1:img_x

Example:
convert_linidx_to_pxlidx(1,10,10) returns [1,1]
convert_linidx_to_pxlidx(10,10,10) returns [1,10]
convert_linidx_to_pxlidx(91,10,10) returns [10,1]
convert_linidx_to_pxlidx(100,10,10) returns [10,10]
}
\description{
Convert linear indices to pixel indices
.
}
