% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{convert_pxlidx_to_linidx}
\alias{convert_pxlidx_to_linidx}
\title{Convert pixel indices to linear indices}
\usage{
convert_pxlidx_to_linidx(row, col, img_x, img_y)
}
\arguments{
\item{row}{the row index of the pixel(s)}

\item{col}{the column index of the pixel(s)}

\item{img_x}{the width of the image in pixels}

\item{img_y}{the height of the image in pixels}
}
\value{
integer indices in 1:(img_x*img_y)

Example:
convert_pxlidx_to_linidx(1,1,10,10) returns 1
convert_pxlidx_to_linidx(2,1,10,10) returns 11
convert_pxlidx_to_linidx(10,1,10,10) returns 91
convert_pxlidx_to_linidx(10,10,10,10) returns 100
}
\description{
Cube layout indices are of the form [row, col].  This function returns the linear index in band sequential layout of [row, col]
}
\details{
row and col can be vectors, in which case they must have the same length, and the output will be this length.
Both row and col are assumed to be 1-indexed (vs. 0), and the output will be returned 1-indexed.
}
