% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{datcub_band_stats}
\alias{datcub_band_stats}
\title{Compute summary statistics over the bands of a data cube}
\usage{
datcub_band_stats(datcub, stat = "mean")
}
\arguments{
\item{datcub}{a data cube of dimensions (nrows, ncols, nbands)}

\item{stat}{the summary statistic to compute, can be one of:
'sum', 'mean', 'sd', 'min', 'max', 'median', 'q25' (0.25 quantile), 'q75', (0.75 quantile)}
}
\value{
a matrix (dimension = nrows x ncols) containing the summary statistic per band
}
\description{
Compute summary statistics over the bands of a data cube
}
\details{
NAs will be ommitted
}
