% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{flip_datcub}
\alias{flip_datcub}
\title{Flip a data cube, along either it's width, height, or bands}
\usage{
flip_datcub(datcub, flipX = F, flipY = F, flipZ = F)
}
\arguments{
\item{datcub}{a data cube}

\item{flipX}{boolean, whether to reverse the column order}

\item{flipY}{boolean, whether to reverse the row order}

\item{flipZ}{boolean, whether to reverse the band order}
}
\value{
a cube of the same
}
\description{
Flip a data cube, along either it's width, height, or bands
}
