% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns_io.R
\name{load_SOM_products}
\alias{load_SOM_products}
\title{Load all SOM-related NeuroScope products}
\usage{
load_SOM_products(
  data_file = NULL,
  include_file = NULL,
  nna_file = NULL,
  wgtcub_file = NULL,
  nunr_file = NULL,
  ctab_file = NULL,
  cadj_file = NULL
)
}
\arguments{
\item{data_file}{filename of the data file}

\item{include_file}{filename of the include file}

\item{nna_file}{filename of the nna file}

\item{wgtcub_file}{filename of the wgtcub file}

\item{nunr_file}{filename of the nunr file}

\item{ctab_file}{filename of the ctab file}

\item{cadj_file}{filename of the cadj file}
}
\value{
a list with components
\itemize{
\item X the data cube, converted to a data matrix (bands in columns)

}
}
\description{
Load all SOM-related NeuroScope products
}
