% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns_io.R
\name{read_ctab}
\alias{read_ctab}
\title{Reading and processing a color table file}
\usage{
read_ctab(filename)
}
\arguments{
\item{filename}{string specifying the file path & name of the color table file}
}
\value{
a data frame with columns:
\itemize{
  \item class:  a character vector defining the class label (see ?read_khoros_includefile for more information on possible class label characters)
  \item R:  an integer vector in [0-255] defining the red plane of the class color
  \item G:  an integer vector in [0-255] defining the green plane of the class color
  \item B:  an integer vector in [0-255] defining the blue plane of the class color
}

Example:
read_ctab('6d-synthetic_30.ctab')
}
\description{
Color tables specify the RGB codes for the classes defined in either the include file or remap.
}
