% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns_io.R
\name{read_incl}
\alias{read_incl}
\title{Reading and processing an include file}
\usage{
read_incl(filename, img_x, img_y)
}
\arguments{
\item{filename}{string specifying the file path & name of the include file}

\item{img_x}{integer, width of image in pixels}

\item{img_y}{integer, height of image in pixels}
}
\value{
a list with components:
\itemize{
  \item masked:  an img_y x img_x boolean matrix whose [i,j] entry specifies whether the pixel in row i, column j of the image should be masked.
  \item class:  an img_y x img_x character matrix whose [i,j] entry defines the class label of the pixel in row i, column j of the image.  Due to Khoros limitations, the available classes are A-Z,a-z plus some other special characters.\cr
                Note:  labeling of all pixels in the include file is not mandatory.  If no class information is found, NA will be returned.
}

Example:
read_incl('8class-caseI-all.incl', 15, 15)
}
\description{
Include files contain information on which parts of the image cube are masked (i.e., pixels that should be ignored) as well as the class designation of each pixel, if available.
}
