% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khoros_io.R
\name{read_khoros_header}
\alias{read_khoros_header}
\title{Read a khoros viff header}
\usage{
read_khoros_header(filename)
}
\arguments{
\item{filename}{string specifying the file path & name of the khoros viff cube}
}
\value{
list with components:
\itemize{
   \item imgEndian = string, either 'little' or 'big'
   \item imgW = integer, width of image in pixels
   \item imgH = integer, height of image in pixels
   \item imgB = integer, number of image bands
   \item imgT = string, image data type ('double', 'integer', etc) \cr
         Note that this data type is for R's \code{readBin} function.
   \item imgTsize = integer, size of data type in bytes
   \item imgTsigned = boolean, TRUE for signed data
}

Example:
read_khoros_header('8class-caseI-n_m0v5.viff')
}
\description{
Information about the image cube is stored in the proprietary Khoros header format.  This function reads that header
and extracts the relevant characteristics needed to read the binary image cube.  Note that this can only parse viff
files (not kdf format; if you have kdf you must convert in Khoros with \code{kformats})
}
