% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khoros_io.R
\name{read_khoros_viff}
\alias{read_khoros_viff}
\title{Import a Khoros cube in viff format}
\usage{
read_khoros_viff(filename, verbose = T)
}
\arguments{
\item{filename}{string specifying the file path & name of the khoros viff cube}

\item{verbose}{whether to print the (1,1) and (img_x,img_y) pixel values to the console}
}
\value{
a 3-dimensional array whose [i,j,k] entry represents the value in row i, column j and band k of the image cube.
}
\description{
The header will be read to obtain the pertinent information required for reading the binary image cube:  image width & height, number of image bands, etc.  The actual reading is done with \code{readBin}.
}
