% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns_io.R
\name{read_nna}
\alias{read_nna}
\title{Read a the min/max mapping stored in a .nna file}
\usage{
read_nna(filename, img_z)
}
\arguments{
\item{filename}{string specifying the file path & name of the nna file}

\item{img_z}{the number of bands in the image that this nna corresponds to}
}
\value{
a list with component:
 \itemize{
  \item input_min, the min of network inputs, by dimension
  \item input_max, the max of network inputs, by dimension
  \item output_min, the min of network outputs
  \item output_max, the max of network outputs
}
}
\description{
Read a the min/max mapping stored in a .nna file
}
