% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns_io.R
\name{read_nunr}
\alias{read_nunr}
\title{Reading and processing a nunr file}
\usage{
read_nunr(filename, som_x, som_y, img_x, img_y)
}
\arguments{
\item{filename}{string specifying the file path & name of the color table file}

\item{som_x}{integer, width of SOM lattice in neurons}

\item{som_y}{integer, height of SOM lattice in neurons}

\item{img_x}{integer, width of image cube in pixels}

\item{img_y}{integer, height of image cube in pixels}
}
\value{
a list with components:
\itemize{
  \item density, a (som_y x som_x) length integer vector whose i-th element contains the number of image pixels (observations) mapped to prototype i.
  \item class, a (som_y x som_y) length character vector whose i-th element contains the majority label of the pixels mapped to prototype i.
  \item mapping, an integer matrix of dimension (img_y, img_x) whose [i,j] entry is the (linear) index of the prototype which pixel [i,j] was mapped to.
  \item som_indexmap, an integer matrix of dimension (som_y, som_x) whose [i,j] entry is the linear index of the neuron at the [i,j] location in the lattice.
}

Special note:
The indexing starts at the SW corner of the lattice, reading along the rows, and finishes at the NE corner of the lattice.
Thus, the first element of the density and class vectors corresponds to the prototype attached to the neuron at the SW corner of the lattice.
View som_indexmap for inspection of this.
The indexing is 1-based.

Example:
read_nunr('8class-caseI.27x27.k3.r.5000000.nunr')
}
\description{
NUNR files contain neuron-specific information about the mapping of the data through the SOM.
}
