% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{vis_classmap}
\alias{vis_classmap}
\title{Visualize a class map (class labels) of image pixels}
\usage{
vis_classmap(class_matrix, ctab, pixel_expansion_factor = 1)
}
\arguments{
\item{class_matrix}{a matrix (nrows = img_y, ncols = img_x) whose [i,j] entry is a character label for pixel[i,j]}

\item{ctab}{a color table data frame, such as the result of calling \code{read_ctab}}

\item{pixel_expansion_factor}{an integer multiple dictating how much to expand each pixel in the resulting image.  No expansion correspoinds to a value of 1.  A value of 2, for example, will double both the width and height of the resulting image.  This is most useful when visualizing information about prototypes since they SOM lattice is generally much smaller than the image data cube.}
}
\value{
none, a visualization

Example:
vis_classmap(read_incl('8class-caseI-all.incl', 15, 15)$class, ctab, 1)
}
\description{
Provides a visual of the class map with colors specified by the color table data frame.
}
