% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{vis_matrix_discrete}
\alias{vis_matrix_discrete}
\title{Visualize discrete values in a matrix}
\usage{
vis_matrix_discrete(mat, colors = NULL, seed = 100)
}
\arguments{
\item{mat}{the matrix whose values will be visualized}

\item{colors}{an option vector containing the colors used during visualization.
The colors will be assigned in ascending order of the unique values in mat, so that
colors[1] corresponds to the value sort(unique(mat))[1],
colors[2] corresponds to the value sort(unique(mat))[2], and so on.
If not given, a color vector will be sampled from R's built in color palette}

\item{seed}{the sampling seed used for setting the colors vector internally.}
}
\value{
nothing, visualization is produced
}
\description{
This is the equivalent of a heatmap for discrete values
}
