% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns_io.R
\name{write_ctab}
\alias{write_ctab}
\title{Write a color table to file}
\usage{
write_ctab(ctab, filename)
}
\arguments{
\item{ctab}{the color table dataframe to write Must have fields  \code{$class}, \code{$R}, \code{$G}, \code{$B}.}

\item{filename}{the file path name to be written to}
}
\description{
Given an color table, write its contents as a text file, with NeuroScope formatting (i.e., "color_table" and "end color_table" tags will be appended).
}
