% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns_io.R
\name{write_incl}
\alias{write_incl}
\title{Writing an include file in ascii format}
\usage{
write_incl(classmap, inclmap, img_x = NULL, img_y = NULL, filename)
}
\arguments{
\item{classmap}{character vector, length = number of pixels in image.  Each element contains a pixel label.}

\item{inclmap}{boolean vector, length = number of pixels in image.  Each element defines whether the pixel is included (=TRUE) or masked (=FALSE) in the image.}

\item{img_x}{integer, width of image cube in pixels}

\item{img_y}{integer, height of image cube in pixels}

\item{filename}{string, address of output file}
}
\value{
None

Special note:  The vectors are assumed to be ordered such that the first element represents the NW pixel and last element represents the SE pixel
}
\description{
Include files are ascii files.
Each row (entry) contains an include statement of the form "include area x0 y0 x1 y1 C"
where (x0,y0) and (x1,y1) are the x-y coordinates of the top left and bottom right (respectively) pixels in the defined region.
C = a class definition.
}
