% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khoros_io.R
\name{write_khoros_header_KRGB}
\alias{write_khoros_header_KRGB}
\title{Write a Khoros KRGB header}
\usage{
write_khoros_header_KRGB(fid, imgE, imgW, imgH, comment = " ")
}
\arguments{
\item{fid}{a connection to a binary file, see \code{?file} for more info.}

\item{imgE}{the file endian-ness, can be 'little' or 'big'}

\item{imgW}{the image width, in pixels}

\item{imgH}{the image height, in pixels}

\item{comment}{optional, any comment string the user wishes to add to the header information. Default = ' ' adds nothing. This must be < 512 characters.}
}
\value{
nothing
}
\description{
If a connection to a binary writeable file is open, this function will write the header required
for Khoros to recognize the cube as KRGB formatted.
}
\details{
This is translated from Khoros I/O manipulation originally written in Matlab by Patrick O'Driscoll.
}
