% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khoros_io.R
\name{write_khoros_KRGB}
\alias{write_khoros_KRGB}
\title{Write a 3-band RGB array as a Khoros KRGB formatted viff cube}
\usage{
write_khoros_KRGB(rgb_img, filename)
}
\arguments{
\item{rgb_img}{a 3-band RGB array, each element in [0-255]}

\item{filename}{full file path to save the resulting .viff}
}
\value{
nothing
}
\description{
Write a 3-band RGB array as a Khoros KRGB formatted viff cube
}
\details{
The image width and height will be inferred from the dimensions of \code{rgb_img}.
A call to \code{file()} is made to open a binary file connection, \code{write_khoros_KRGB_header} is called to write the
appropriate header, and then the integer RGB codes are written.
}
