% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khoros_io.R
\name{write_khoros_viff}
\alias{write_khoros_viff}
\title{Writing a Khoros data cube in .viff format}
\usage{
write_khoros_viff(
  X,
  img_x = NULL,
  img_y = NULL,
  filename,
  header_comment = " "
)
}
\arguments{
\item{X}{either a data matrix (nrows = img_x * img_y, ncols = img_z), or a data cube (nrows = img_y, ncols = img_x, nbands = img_z)}

\item{img_x}{integer, width of image cube in pixels}

\item{img_y}{integer, height of image cube in pixels}

\item{filename}{string, address of output file}

\item{header_comment}{string, any comments about the data to include in the header}
}
\value{
None

Special note:
}
\description{
Khoros data cubes contain a 1024 byte header, followed by the data cube in band sequential layout
}
\details{
If X is given as a data matrix then img_x and img_y must also be provided. Otherwise, if X is a data cube,
these are taken from cube dimensions.

If X is a data matrix it is assumed to be ordered such that the first row represents the (1,1) pixel and last row represents the (img_x,img_y) pixel
}
