% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns_io.R
\name{write_nna}
\alias{write_nna}
\title{Write a nna (minmax table)}
\usage{
write_nna(
  filename,
  input_min,
  input_max,
  output_min,
  output_max,
  group_size = 5
)
}
\arguments{
\item{filename}{string specifying the file path & name of the nna file}

\item{input_min}{vector of min range of network inputs (by dim)}

\item{input_max}{vector of max range of network inputs (by dim)}

\item{output_min}{vector of min range of network outputs}

\item{output_max}{vector of max range of network outputs}

\item{group_size}{size of groups for each line of the nna file}
}
\description{
Write a nna (minmax table)
}
